LibWeb: Construct filter chain at paint time
This allows us to use style and box metrics when applying filters.
