+++ Libraries/LibWeb/CSS/ComputedValues.h
@@ -156,8 +156,8 @@ public:
     static CSS::Display display() { return CSS::Display { CSS::DisplayOutside::Inline, CSS::DisplayInside::Flow }; }
     static Color color() { return Color::Black; }
     static Color stop_color() { return Color::Black; }
-    static Optional<Gfx::Filter> backdrop_filter() { return {}; }
-    static Optional<Gfx::Filter> filter() { return {}; }
+    static CSS::Filter backdrop_filter() { return CSS::Filter::make_none(); }
+    static CSS::Filter filter() { return CSS::Filter::make_none(); }
     static Color background_color() { return Color::Transparent; }
     static CSS::ListStyleType list_style_type() { return CSS::CounterStyleNameKeyword::Disc; }
     static CSS::ListStylePosition list_style_position() { return CSS::ListStylePosition::Outside; }
@@ -504,8 +504,8 @@ public:
     CSS::JustifyContent justify_content() const { return m_noninherited.justify_content; }
     CSS::JustifySelf justify_self() const { return m_noninherited.justify_self; }
     CSS::JustifyItems justify_items() const { return m_noninherited.justify_items; }
-    Optional<Gfx::Filter> const& backdrop_filter() const { return m_noninherited.backdrop_filter; }
-    Optional<Gfx::Filter> const& filter() const { return m_noninherited.filter; }
+    CSS::Filter const& backdrop_filter() const { return m_noninherited.backdrop_filter; }
+    CSS::Filter const& filter() const { return m_noninherited.filter; }
     Vector<ShadowData> const& box_shadow() const { return m_noninherited.box_shadow; }
     CSS::BoxSizing box_sizing() const { return m_noninherited.box_sizing; }
     CSS::Size const& width() const { return m_noninherited.width; }
@@ -749,8 +749,8 @@ protected:
         CSS::LengthBox inset { InitialValues::inset() };
         CSS::LengthBox margin { InitialValues::margin() };
         CSS::LengthBox padding { InitialValues::padding() };
-        Optional<Gfx::Filter> backdrop_filter { InitialValues::backdrop_filter() };
-        Optional<Gfx::Filter> filter { InitialValues::filter() };
+        CSS::Filter backdrop_filter { InitialValues::backdrop_filter() };
+        CSS::Filter filter { InitialValues::filter() };
         BorderData border_left;
         BorderData border_top;
         BorderData border_right;
@@ -928,8 +928,8 @@ public:
     void set_list_style_type(CSS::ListStyleType value) { m_inherited.list_style_type = value; }
     void set_list_style_position(CSS::ListStylePosition value) { m_inherited.list_style_position = value; }
     void set_display(CSS::Display value) { m_noninherited.display = value; }
-    void set_backdrop_filter(Optional<Gfx::Filter> backdrop_filter) { m_noninherited.backdrop_filter = move(backdrop_filter); }
-    void set_filter(Optional<Gfx::Filter> filter) { m_noninherited.filter = move(filter); }
+    void set_backdrop_filter(CSS::Filter const& backdrop_filter) { m_noninherited.backdrop_filter = backdrop_filter; }
+    void set_filter(CSS::Filter const& filter) { m_noninherited.filter = filter; }
     void set_border_bottom_left_radius(CSS::BorderRadiusData value)
     {
         m_noninherited.has_noninitial_border_radii = true;
