+++ Libraries/LibWeb/Painting/PaintableBox.cpp
@@ -30,6 +30,7 @@
 #include <LibWeb/Painting/TextPaintable.h>
 #include <LibWeb/Painting/ViewportPaintable.h>
 #include <LibWeb/Platform/FontPlugin.h>
+#include <LibWeb/SVG/SVGFilterElement.h>
 #include <LibWeb/Selection/Selection.h>
 
 namespace Web::Painting {
@@ -582,14 +583,15 @@ void PaintableBox::paint_border(DisplayListRecordingContext& context) const
 void PaintableBox::paint_backdrop_filter(DisplayListRecordingContext& context) const
 {
     auto const& backdrop_filter = computed_values().backdrop_filter();
-    if (!backdrop_filter.has_value()) {
+    if (backdrop_filter.is_none()) {
         return;
     }
 
     auto backdrop_region = context.rounded_device_rect(absolute_border_box_rect());
     auto border_radii_data = normalized_border_radii_data();
     ScopedCornerRadiusClip corner_clipper { context, backdrop_region, border_radii_data };
-    context.display_list_recorder().apply_backdrop_filter(backdrop_region.to_type<int>(), border_radii_data, backdrop_filter.value());
+    if (auto resolved_backdrop_filter = resolve_filter(backdrop_filter); resolved_backdrop_filter.has_value())
+        context.display_list_recorder().apply_backdrop_filter(backdrop_region.to_type<int>(), border_radii_data, *resolved_backdrop_filter);
 }
 
 void PaintableBox::paint_background(DisplayListRecordingContext& context) const
@@ -1629,6 +1631,82 @@ PaintableBox const* PaintableBox::nearest_scrollable_ancestor() const
         paintable = paintable->containing_block();
     }
     return nullptr;
+}
+
+Optional<Gfx::Filter> PaintableBox::resolve_filter(CSS::Filter const& computed_filter) const
+{
+    Optional<Gfx::Filter> resolved_filter;
+    for (auto const& filter : computed_filter.filters()) {
+        filter.visit(
+            [&](CSS::FilterOperation::Blur const& blur) {
+                auto resolved_radius = blur.resolved_radius(layout_node_with_style_and_box_metrics());
+                auto new_filter = Gfx::Filter::blur(resolved_radius, resolved_radius);
+
+                resolved_filter = resolved_filter.has_value()
+                    ? Gfx::Filter::compose(new_filter, *resolved_filter)
+                    : new_filter;
+            },
+            [&](CSS::FilterOperation::DropShadow const& drop_shadow) {
+                CSS::CalculationResolutionContext context {
+                    .length_resolution_context = CSS::Length::ResolutionContext::for_layout_node(layout_node_with_style_and_box_metrics()),
+                };
+                auto to_px = [&](CSS::LengthOrCalculated const& length) {
+                    return static_cast<float>(length.resolved(context).map([&](auto& it) { return it.to_px(layout_node_with_style_and_box_metrics()).to_double(); }).value_or(0.0));
+                };
+                // The default value for omitted values is missing length values set to 0
+                // and the missing used color is taken from the color property.
+                auto new_filter = Gfx::Filter::drop_shadow(to_px(drop_shadow.offset_x),
+                    to_px(drop_shadow.offset_y),
+                    drop_shadow.radius.has_value() ? to_px(*drop_shadow.radius) : 0.0f, drop_shadow.color.has_value() ? *drop_shadow.color : this->computed_values().color());
+
+                resolved_filter = resolved_filter.has_value()
+                    ? Gfx::Filter::compose(new_filter, *resolved_filter)
+                    : new_filter;
+            },
+            [&](CSS::FilterOperation::Color const& color_operation) {
+                auto new_filter = Gfx::Filter::color(color_operation.operation, color_operation.resolved_amount());
+
+                resolved_filter = resolved_filter.has_value()
+                    ? Gfx::Filter::compose(new_filter, *resolved_filter)
+                    : new_filter;
+            },
+            [&](CSS::FilterOperation::HueRotate const& hue_rotate) {
+                auto new_filter = Gfx::Filter::hue_rotate(hue_rotate.angle_degrees(layout_node_with_style_and_box_metrics()));
+
+                resolved_filter = resolved_filter.has_value()
+                    ? Gfx::Filter::compose(new_filter, *resolved_filter)
+                    : new_filter;
+            },
+            [&](CSS::URL const& css_url) {
+                auto& url_string = css_url.url();
+
+                if (url_string.is_empty() || !url_string.starts_with('#'))
+                    return;
+
+                auto fragment_or_error = url_string.substring_from_byte_offset(1);
+
+                if (fragment_or_error.is_error())
+                    return;
+
+                // FIXME: Support urls that are not only composed of a fragment.
+                auto maybe_filter = document().get_element_by_id(fragment_or_error.value());
+
+                if (!maybe_filter)
+                    return;
+
+                if (auto* filter_element = as_if<SVG::SVGFilterElement>(*maybe_filter)) {
+                    auto new_filter = filter_element->gfx_filter();
+
+                    if (!new_filter.has_value())
+                        return;
+
+                    resolved_filter = resolved_filter.has_value()
+                        ? Gfx::Filter::compose(*new_filter, *resolved_filter)
+                        : new_filter;
+                }
+            });
+    }
+    return resolved_filter;
 }
 
 }
