+++ Libraries/LibWeb/Painting/PaintableBox.h
@@ -263,6 +263,8 @@ public:
     [[nodiscard]] RefPtr<ClipFrame const> enclosing_clip_frame() const { return m_enclosing_clip_frame; }
     [[nodiscard]] RefPtr<ClipFrame const> own_clip_frame() const { return m_own_clip_frame; }
 
+    Optional<Gfx::Filter> resolve_filter(CSS::Filter const& computed_filter) const;
+
 protected:
     explicit PaintableBox(Layout::Box const&);
     explicit PaintableBox(Layout::InlineNode const&);
