+++ Libraries/LibWeb/Painting/SVGSVGPaintable.cpp
@@ -56,8 +56,12 @@ void SVGSVGPaintable::paint_svg_box(DisplayListRecordingContext& context, Painta
         context.display_list_recorder().apply_opacity(computed_values.opacity());
     }
 
-    if (filter.has_value()) {
-        context.display_list_recorder().apply_filter(filter.value());
+    auto filter_applied = false;
+    if (filter.has_filters()) {
+        if (auto resolved_filter = svg_box.resolve_filter(filter); resolved_filter.has_value()) {
+            context.display_list_recorder().apply_filter(*resolved_filter);
+            filter_applied = true;
+        }
     }
 
     if (compositing_and_blending_operator != Gfx::CompositingAndBlendingOperator::Normal) {
@@ -94,7 +98,7 @@ void SVGSVGPaintable::paint_svg_box(DisplayListRecordingContext& context, Painta
         context.display_list_recorder().restore();
     }
 
-    if (filter.has_value()) {
+    if (filter_applied) {
         context.display_list_recorder().restore();
     }
